Separate rendering tasks in Game into GameRender
