+++ Game.java
@@ -10,8 +10,10 @@ import java.util.concurrent.TimeUnit;
 import javax.imageio.*;
 import javax.swing.*;
 
-public class Game extends JPanel {
+public class Game {
     private Timer timer;
+    private GameRender gameRender;
+    //renderable game objects
     private Snake snake;
     private Point cherry;
     private Point extraLife;
@@ -19,11 +21,12 @@ public class Game extends JPanel {
     private int[] best = new int[10]; // best score is last
     private int extraLives =0;
     private int newCherries =0;
-    private BufferedImage image;
-    private BufferedImage extraLifeImage;
+
+
     private GameStatus status;
-    private boolean didLoadCherryImage = true;
-    private boolean didLoadExtraLife = true;
+
+
+
 
 
     private static Font FONT_M = new Font("ArcadeClassic", Font.PLAIN, 24);
@@ -33,38 +36,55 @@ public class Game extends JPanel {
     private static int WIDTH = 760;
     private static int HEIGHT = 520;
     private static int DELAY = 50;
+    //private static Game _instance = null;
 
     public Game() {
-        try {
-            image = ImageIO.read(new File("cherry.png"));
-        } catch (IOException e) {
-            didLoadCherryImage = false;
-        }
-
-        try{
-            extraLifeImage = ImageIO.read(new File(".png"));
-        }catch(IOException e){
-            didLoadExtraLife = false;
-        }
-        addKeyListener(new KeyListener());
-        setFocusable(true);
-        setBackground(Color.black);
-        setDoubleBuffered(true);
-
+        gameRender = new GameRender(this, new KeyListener());
+        //TODO game objects
         snake = new Snake(WIDTH / 2, HEIGHT / 2);
         status = GameStatus.NOT_STARTED;
-        repaint();
+    }
+/*
+    public static Game getInstance() {
+        if(_instance == null) {
+            _instance = new Game();
+        }
+        return _instance;
     }
 
-    @Override
-    public void paintComponent(Graphics g) {
-        super.paintComponent(g);
+ */
 
-        render(g);
-
-        Toolkit.getDefaultToolkit().sync();
+    public GameRender getGameRender() {
+        return gameRender;
     }
 
+    public Snake getSnake() {
+        return snake;
+    }
+
+    public Point getCherry() {
+        return cherry;
+    }
+
+    public Point getExtraLife() {
+        return extraLife;
+    }
+    public int getPoints() {
+        return points;
+    }
+    public int getExtraLives() {
+        return  extraLives;
+    }
+
+    public GameStatus getStatus() {
+        return status;
+    }
+
+    public int[] getBest() {
+        return best;
+    }
+
+    //TODO Game logic
     private void update() {
         snake.move();
 
@@ -93,6 +113,7 @@ public class Game extends JPanel {
         checkForGameOver();
     }
 
+    //TODO Game logic
     private void reset() {
         points = 0;
         cherry = null;
@@ -100,7 +121,7 @@ public class Game extends JPanel {
         snake = new Snake(WIDTH / 2, HEIGHT / 2);
         setStatus(GameStatus.RUNNING);
     }
-
+    //TODO Game logic
     private void extraLife(){
         setStatus(GameStatus.EXTRA_LIFE);
         cherry = null;
@@ -112,7 +133,7 @@ public class Game extends JPanel {
         //}
         setStatus(GameStatus.RUNNING);
     }
-
+    //TODO Game logic
     private void setStatus(GameStatus newStatus) {
         switch(newStatus) {
             case EXTRA_LIFE:
@@ -143,10 +164,11 @@ public class Game extends JPanel {
         status = newStatus;
     }
 
+    //TODO Game logic
     private void togglePause() {
         setStatus(status == GameStatus.PAUSED ? GameStatus.RUNNING : GameStatus.PAUSED);
     }
-
+    //TODO Game logic
     private void checkForGameOver() {
 
         Point head = snake.getHead();
@@ -172,104 +194,20 @@ public class Game extends JPanel {
         }
 
     }
-
-    public void drawCenteredString(Graphics g, String text, Font font, int y) {
-        FontMetrics metrics = g.getFontMetrics(font);
-        int x = (WIDTH - metrics.stringWidth(text)) / 2;
-
-        g.setFont(font);
-        g.drawString(text, x, y);
-    }
-
-    private void render(Graphics g) {
-        Graphics2D g2d = (Graphics2D) g;
-
-        g2d.setColor(new Color(53, 220, 8));
-        g2d.setFont(FONT_M);
-
-        if (status == GameStatus.NOT_STARTED) {
-            drawCenteredString(g2d, "SNAKE", FONT_XL, 200);
-            drawCenteredString(g2d, "GAME", FONT_XL, 300);
-            drawCenteredString(g2d, "Press  any  key  to  begin ", FONT_M_ITALIC, 330);
-
-            return;
-        }
-
-        Point p = snake.getHead();
-
-        g2d.drawString("SCORE: " + String.format ("%04d", points), 20, 30);
-        g2d.drawString("EXTRA LIVES: " + extraLives,320,30);
-        g2d.drawString("BEST: " + String.format ("%04d", best[best.length-1]), 660, 30);
-
-        if (cherry != null) {
-            if (didLoadCherryImage) {
-                g2d.drawImage(image, cherry.getX(), cherry.getY(), 60, 60, null);
-            } else {
-                g2d.setColor(Color.RED);
-                g2d.fillOval(cherry.getX(), cherry.getY(), 10, 10);
-                g2d.setColor(new Color(53, 220, 8));
-            }
-        }
-
-        if(extraLife!=null){
-            if(didLoadExtraLife){
-                g2d.drawImage(extraLifeImage,extraLife.getX(),extraLife.getY(),60,60,null);
-            }else {
-                g2d.setColor(Color.GREEN);
-                g2d.fillOval(extraLife.getX(),extraLife.getY(),10,10);
-                g2d.setColor(new Color(53,220,8));
-            }
-        }
-
-
-
-        if (status == GameStatus.GAME_OVER) {
-            renderEndGame(g2d);
-
-        }
-
-        if (status == GameStatus.PAUSED) {
-        	drawCenteredString(g2d, "Paused", FONT_L, 300);
-        	drawCenteredString(g2d, "Press R to restart", FONT_M_ITALIC, 330);
-        }
-
-        g2d.setColor(new Color(245, 3, 52));
-        g2d.fillRect(p.getX(), p.getY(), 10, 10);
-        g2d.setColor(new Color(74, 245, 14));
-        g2d.fillRect(p.getX(), p.getY(), 10, 10);
-
-        for(int i = 0, size = snake.getTail().size(); i < size; i++) {
-            Point t = snake.getTail().get(i);
-            g2d.setColor(new Color(71, 128, 0));
-            g2d.fillRect(t.getX(), t.getY(), 10, 10);
-        }
-
-        g2d.setColor(new Color(71, 128, 0));
-        g2d.setStroke(new BasicStroke(4));
-        g2d.drawRect(20, 40, WIDTH, HEIGHT);
-    }
-    public void renderEndGame(Graphics2D g2d) {
-        drawCenteredString(g2d, "Press  enter  to  start  again ", FONT_M_ITALIC, 140);
-        drawCenteredString(g2d, "GAME OVER", FONT_L, 110);
-        drawCenteredString(g2d, "Best Scores:", FONT_M, 170);
-        for(int i = best.length-1; i >=0 ; i--) {
-            String score = (best.length - i) + ". "  + best[i];
-            drawCenteredString(g2d, score, FONT_M, 425 - i*25);
-        }
-    }
-
+    //TODO game logic
     public void spawnCherry() {
         cherry = new Point((new Random()).nextInt(WIDTH - 60) + 20,
                 (new Random()).nextInt(HEIGHT - 60) + 40);
     }
 
+    //TODO game logic
     public void spawnExtraLife(){
 
         extraLife = new Point((new Random()).nextInt(WIDTH - 60) + 20,
                 (new Random()).nextInt(HEIGHT - 60) + 40);
 
     }
-
+    //TODO game mixed with Rendering but correctly (asks for render)
     private class KeyListener extends KeyAdapter {
         @Override
         public void keyPressed(KeyEvent e) {
@@ -283,34 +221,23 @@ public class Game extends JPanel {
                     case KeyEvent.VK_DOWN: snake.turn(Direction.DOWN); break;
                 }
             }
-
-            if (status == GameStatus.NOT_STARTED) {
+            if (status == GameStatus.NOT_STARTED)
                 setStatus(GameStatus.RUNNING);
-            }
-
-            if (status == GameStatus.GAME_OVER && key == KeyEvent.VK_ENTER) {
+            if (status == GameStatus.GAME_OVER && key == KeyEvent.VK_ENTER)
                 reset();
-            }
-
-          //Reset
-            if(key == KeyEvent.VK_R && status == GameStatus.PAUSED) {
-            	 reset();
-             	 
-            }
-
+            if(key == KeyEvent.VK_R && status == GameStatus.PAUSED)
+                reset();
             if (key == KeyEvent.VK_P) {
                 togglePause();
-                repaint();
+                gameRender.repaint();
             }
         }
     }
-
+    //TODO game mixed with Rendering but correctly (asks for render)
     private class GameLoop extends java.util.TimerTask {
         public void run() {
             update();
-            repaint();
+            gameRender.repaint();
         }
     }
-
-
 }
